upgraded termination checker to handle mutual recursion / corecursion
Each arc in the call graph now carries a flag whether it is a lazy
(corecursive) call or not.  When combining calls, lazy is dominant.
In the end, when we check in the completed call graph whether all
idempotent calls contain a decreasing entry in the diagonal, the
"guardedness" item is taken into consideration for all lazy calls.

Test example is: test/succeed/StreamProcEat.agda  
(Hancock et al)