Add support for mixfix lambdas.
The most unexpected change is that expression can now contain
identifiers that are unbound in the scope of the expression (because
they are bound within). So the checks must be relaxed. The error
messages do not seem to suffer however, contrary to a comment in
parseApplication.

Something I am unsure about in this patch:

-        isLocal    = case use of
-            UseBoundNames     -> \x -> Set.member x set
-            DontUseBoundNames -> \x -> Set.member x connames || not (Set.member x conparts)
+        isLocal x = not (Set.member x allParts)
+        -- If string is a part of notation, it cannot be used as an identifier.

However everything seems to be working.