Added .mailmap file.
In the migration from Darcs to GitHub, we lost some information about
the authors. The information from Git commands,
e.g. git shortlog -nse, is better by using this file. The creation of
this file was based on the .authorspellings Darcs file.

Note: In the Darcs repository, Nils Anders Danielsson and Dominique
Devriese had 1479 and 73 patches, respectively; but in the master
branch of the GitHub repository, they only have 1471 and 72 commits,
respectively. I couldn't fix it.