Unquoting extended lambdas.
The way unquoting works is a little awkward at the moment. We
unquote to internal syntax, and then we convert the result back
to abstract syntax for type checking. To get extended lambdas
working I had to add them to the internal syntax. They shouldn't
leak out of the unquote -> reify -> type check chain though.