[ issue892 ] fixed
Required a slight change in the semantics of module application.
Previously re-exported functions were not redefined in the new module,
but simply aliased by the scope checker. This is wrong as issue 892
shows, so now all exported functions are redefined, which may lead to
ambiguous identifier errors in code that previously checked.