Lensify TCState.
Doing this now because I want to split TCState without committing on the
exact structure forever.  Specifically, I've split the state in what is
used up to scope checking and what is used after -- we need this to
implement incremental type checking nicely.

We should probably lensify stPermanentState as well, and kill state
modifying functions (since they're now useless), of which the code base
is generously peppered, but I figured it's better to have a "compact"
diff for now.