[ Issue 1428 ] New sort SizeUniv, but unused yet.
SizeUniv will be the sort of builtin types SIZE and SIZELT.
Its purpose is to prevent building functions on Size, especially
(pred : (i : Size) -> Size< i) which triggers issue 1428.

So far, it is unused, as a new Sort is incompatible with the
current (hackish) handling of sort metas, which assumes that every sort
is of the form Set x for some level x (including infinity to represent
sort Inf).  To activate SizeUniv, we need proper sort metas first.