[ Issue 1109, Issue 1436 ] New default precedence level: unrelated.
Note that I decided to use the new default precedence level for every
operator without a fixity declaration, not only those discussed in
connection with Issue 1436. Thus this is not purely a bug-fix. This
commit also relies on the new operator parser and breaks lots of
existing code, so it should not be ported to the maint-2.4.2 branch.

The commit makes operator parsing roughly 25% slower for the standard
library (with my current setup).