[ issue 1396 ] fixed scope issue with open public of parameterised modules
The scope checker computed a renaming mapping old name to new name for
module applications, with each old abstract name mapping to a unique new name.
The problem is that with open public the same old name might need to be copied
several times. For instance

  module A (X : Set) where
    module B (Y : Set) where
      f : Z
    open B public

In this case A.B.f is accessible also as A.f, so if we apply

  module AQ = A Q

we should get both AQ.f and AQ.B.f as copies of A.B.f. Fixed by making renamings
admit duplicate old names.