[ Issue 1428 ] Putting Size and Size< in SizeUniv.
This somewhat works, but redesign of Level is needed to properly account
for the PTS rule (SizeUniv,s,s).  Previously, the only PTS rule
was (s,s',max(s,s')) so the functional PTS rule was just the least upper
bound.  The syntax of levels (sorts) has been optimized for this PTS rule,
currently only the maximum of sorts is handled.  TODO!