[ uhc ] Fix UHC pragma parsing when UHC backend is disabled.
We now store COMPILED_CORE and COMPILED_CORE_DATA pragmas as strings
in the *.agdai files. This avoids having a dependency on uhc-light
for the *.agdai files.

If the UHC backend is disabled, the pragmas are not checked for correctness,
but are stored in the *.agdai files anyway. If the UHC backend is enabled,
they are checked during type checking.