[ exec-tests ] Better fallback code for old process versions.
The old fallback code did manipulate the current directory of the process,
to make sure the test setup code was executed in the correct places.
This breaks down if multiple tests are executed in parallel, as the
current directory is per-process state, and not per-thread.
Instead, just set the current directory for the spawned child
process, but not the current one.
Console output can still be a bit ugly due to the interleaved
output, but this shouldn't cause any problems and only happens
with process older than 1.2.3.