[ LaTeX ] Using the characters' real width
Previously we were moving the column by the length of the text
rather than its size meaning that wide characters in the editor
(e.g. 𝓔 or 𝓜 ) were considered to be of size 1 rather than 2.
As a consequence, it was impossible to have a neat alignment in
both the source code and the resulting pdf. This patch makes sure
we are using the right width instead.