Revert "[ LaTeX ] Using the characters' real width"
This reverts commit fc2df9eea00221b070492195b3edc736ce8b3530.

This commit broke the build when older versions of text that did not
export Data.Text.Unsafe were used. The problem could be fixed by
revising the version bounds for text. However, I would be very
surprised if the number of 16-bit words in a UTF-16-encoded string is
always equal to the number of columns used to render the string with
an arbitrary monospace font, so I have decided to revert the commit.