[ Issue 1592 ] optimized Positivity.setArgs to work for all mutual defs simultaneously.
Removed a source of quadratic behavior (arity computation from graph),
which is most probably an inessential optimization. But it does not obfuscate the code.

The arity computation also only considers the ArgNodes with outgoing
edges now, which fixes the originally reported issue 1592 (though that
is not the end of the story).