[ Issue1366b ] Fixed by including abstract local defs in mutual injectivity check.
Changed semantics of abstract such that local definitions in a where
block are always transparent from the perspective of the parent
function.  I would have guessed this was intended from the beginning.

At least, abstract should satisfy the following property, or?
Have a self-contained module, put everything in a big abstract block, it
should still type-check.  Issue 1366b was one instance where this
invariant was violated.