[ Serialize ] Remove A.ScopedExpr when serializing.
A.Expr are only serialized for dot patterns in pattern synonyms.
The scope should not be needed there.

Anyway, serializing embedded scopes in expressions could be super
costly, so this is out of question.  Thus, if some parts of the scope
should be needed, another solution has to be found.