[ #1108 ] Made ranges strict. Replaced interval lists with vectors.
Ranges are now abstract, and represented as strict vectors of
intervals. This makes Agda much slower on Issue1108-2.agda. I tried
replacing fuseRanges with const (or perhaps flip const), and this made
Agda much faster, so it might be worthwhile investigating some
representation of ranges which enables an efficient implementation of
fuseRanges.