[ #1108 ] Made Range fields lazy.
This is much better than before (for Issue1108-2.agda), and comparable
to replacing the implementation of fuseRanges with const/flip const:

Total                912ms
Miscellaneous          0ms
Parsing               48ms
Parsing.Operators    180ms
Import                 0ms
Deserialization        4ms
Scoping              140ms
Typing               380ms
Typing.OccursCheck     0ms
Termination            0ms
Termination.RecCheck  12ms
Positivity             8ms
Injectivity            0ms
ProjectionLikeness     0ms
Coverage               0ms
Highlighting          28ms
Serialization         52ms
ModuleName            60ms

Now the time needed to type-check Issue1108-2.agda seems to be roughly
linear in the length of the list. If the unsolved meta-variable at the
end is removed, so that an interface file is produced, then the time
required is possibly quadratic in the length of the list. For
Issue1108.agda the time needed to serialise seems to be roughly
linear, but the operator parser might be quadratic, or worse.

The unsolved meta-variables have now been removed, so that
serialisation is performed.

For the standard library the difference seems to be small (compared to
using lists in the definition of ranges).

I tried removing the special cases from fuseRanges, but this was much
worse for (the new) Issue1108-2.agda. I also tried removing the
special cases from the recursive calls, only using them at the
top-level, and this seems to have been slightly worse.