[ deadcode ] Eliminate unused definitions before serialisation
A definition is used if it's exported from the module or mentioned by a
used definition. A side effect of this is that private names can no
longer be in scope at the emacs top-level, since they might have been
removed in the dead code elimination.

This saves over 20% on the serialisation time of the standard library.