Keep track of std-lib as a git submodule
A git submodule is a reference to another git repository from
a parent repository. The reference is a specific directory in the
repository, where the submodule `lives`.

In our case, the parent is agda/agda, and the submodule is
agda/agda-stdlib.  The submodule lives in the directory /std-lib .

Each commit A in the parent repository references a specific commit B of the
child repository. In our case, the commit referenced should be the version
of the standard library that is supposed to work with the contents of
the parent repository.

The parent repository also contains a `.gitmodules` file in the root,
which specifies a URL and branch name for the submodule.

To initalize and update the submodule, you may use the usual `std-lib`
and `up-to-date-stdlib` rules in the Makefile.

After updating the standard library, you may run …

    git add std-lib
    git commit

… from the root to make the repository point to the new version.

Finally, note that the std-lib branch is now specified in the
`.gitmodules` file in the root, and not in the Makefile.

If you change the branch, run …

    git submodule sync

… afterwards to update the contents of the `std-lib` directory to
those of the new branch.