Fixed printing of underscore patterns in errors.
The bug was introduced by the conflation of A.WildP with A.ImplicitP.
Previously, removal of ImplicitPs before printing was fine, as they
had been inserted by the type checker.  But removal of WildPs is not
fine, of course.

Now, nothing is removed, as previous ImplicitPs should be hidden WildPs,
so they are not printed anyway.