[ Emacs ] Partial fix for #1061.
If the information buffer is already displayed, but only in another
frame, then the buffer will hopefully not be displayed in the current
frame. (With the default configuration, display-buffer is very
configurable.)

However, if the information buffer is not already displayed, then it is
likely to be displayed in the current frame, even if Agda was not
invoked from this frame.

I don't think the fix has any effect if the Emacs version is less than
24, and the fix has not been tested using Aquamacs.