[ #1901 ] Changed the definition of OccursWhere.
OccursWhere was defined as a kind of specialised list of elements of a
certain kind, followed by one of two nil constructors. However, one of
the nil constructors (Unknown) always appeared on its own. Now
OccursWhere is either Unknown, or a sequence (Data.Sequence, because the
sequences are sometimes appended to each other) of elements.

This change seems to have little impact on the time required to
type-check the standard library. For benchmark/misc/SlowOccurrences.agda
the heap profile becomes flatter (the peak shown in the first heap
profile in the issue report disappears), but the type-checking time is
basically unchanged. Andreas Nuyts' example also uses less memory.

Note that this commit does not directly address the problem that the
list of OccursWhere elements can contain lots of duplicated information.