[ Fixed #1901 ] Replaced quadratic code with linear.
(In the size of the term. Assuming strict evaluation.)

This commit leads to a speedup of ∼30% for
benchmark/misc/SlowOccurrences.agda (with my present setup). For the
complete sequence of optimisations related to #1901 the speedup is ∼60%,
the time spent on positivity checking according to -vprofile:7 has gone
down from ∼2s to ∼10ms, and the maximum residency is now less than 10%
of what it used to be.

This commit also makes positivity checking ∼70%-80% faster for Andreas
Nuyts' code (with the same caveats as above).