[ #1904 ] Only one file name per Range.
Rather than one file name for every position in every interval in the
Range.

This change solves the problem with "Issue1108.agda with a
snoc-list" (see Issue1904.agda). However, fuseRanges is still
potentially expensive, and the size of the ranges in an expression can
still be quadratic in the size of the expression tree, at least if
sharing is ignored.