Unfixing #1428.  SizeUniv deactivated.
With SizeUniv, sorts are no longer linear order, and the ptsRule is no
longer just maximum (lub).  This could be implemented properly, but is
serious effort due to the optimizations done for levels.

This patch defines SizeUniv = Set, so all the code for SizeUniv should
be dead.  I keep it, because I might come back and implement the ptsRule
properly.