Optimisation: Removed killRange call. Most ranges are not serialised.
This reduces the amount of "void" memory in the biographical heap
profile for benchmark/categories/Language3.agda, and the memory usage
goes down considerably. Agda also seems to type-check the standard
library a little faster.

Note that we no longer kill all definition site ranges in the signature
when creating an interface.

Developed together with Ulf Norell.