Optimisation: Signatures are not instantiated.
The meta-variable instantiations are serialised, and this serialisation
is memoised. The instantiations are applied when the signature is
deserialised. The serialised signature is immediately deserialised, to
ensure that there are no meta-variables in it.

Test case: benchmark/categories/Language3.agda.

TODO: This test case type checks a bit faster. However, it seems as if
it takes roughly 10% more time to type-check the standard library.
Perhaps it would be better to avoid the deserialisation step (and thus
also preserve sharing of meta-variable instantiations).

Developed together with Ulf Norell.