Re #1950 fixed regression, but still work TODO
The displayform is now not rejected.  But still the result is
unparseable, as the additionally introduced arguments are printed
as with patterns.  We currently have no syntax for additional arguments
or copatterns in with-clauses.

One way to go for a quick fix is to ensure that case splitting does not
add trailing arguments which would confuse the printer.

However, in the long run we want to have additional arguments or
copatterns in with-clauses.