Made it possible to print operator grammars.
Using -vscope.grammar:10.

The output tends to contain lots of parentheses. I guess that it would
be easy to make it more pretty.

For the standard library there seems to be little difference in
performance, but this change seems to make Agda roughly 30% slower for
test/Succeed/Issue1108.agda, even if no grammars are printed. For this
reason the printing facility is only available if debugging is
activated (one can for instance give the -fdebug flag to cabal install).

Related issue: #1977.