Clarify for #2018: removed recConType
This field was only used twice but had to be maintained in Substitute
etc.  It contained only redundant information, as the type of the record
constructor can be obtained from the signature via the record
constructor name (as it is done for data types).

Also, the hiding of the parameters of this type was not correctly set by
addConstant.  Thus, it was in fact redundant inconsistent information I
removed.