Fixed #2076, also more sound solution for #2047
With parameter refinement, the "parameter" patterns of an extended
lambda can now be different from variable patterns.  If we just drop
them (plus the associated arguments to the extended lambda), we produce
something

* that violates internal invariants.  In particular, the permutation
  dbPatPerm from the patterns to the telescope can no longer be
  computed.  (And in fact, dropping from the start of the telescope is
  just plainly unsound then.)

* prints the wrong thing (old fix for #2047)

What we do now, is more sound, although not entirely satisfying:
When the "parameter" patterns of an external lambdas are not variable
patterns, we fall back to printing the internal function created for the
extended lambda, instead trying to construct the nice syntax.