Fixed #2076 (for good?) apply an extended lambda to all parameters when reifying
The getDefFreeVars are irrelevant for extended lambdas,
since we print their body, and not their internal name.  Thus, we need
to substitute the parameters into the body.
(We add the whole section in which the extended lambda was
defined to its special parameters, to get the full number of parameters.)