Fixed #2101: declarations in where-block are now private
This has the effect that their type signatures can be checked with
unfolding of abstract definitions.   In particular, we should be able to
move a definition with a where-block into an abstract block (see
enclosed test cases).

As a side effect, two error messages have changed: the top level module
name is now included in the out-of-scope qualified identifier:  see

  test/Fail/PatternSynonymsErrorLocation*.err

However, this can be observed in other error messages as
well (e.g. termination errors) and cries for a systematic solution.
Does not have to be addressed here.