Implemented a git bisect wrapper script.
I used Paolo Capriotti's library optparse-applicative to parse command
line options. The library is based on parser combinators, so it is for
instance easy to express that two flags are mutually exclusive. It is
also easy to support Bash completion, which I think we should add
support for in Agda.