Re #2231 in abstract blocks make #-generated functions abstract
This is in preparation for issue #2231.

Before, they were concrete, since they were anyway not directly
referenceable, thus, it did not matter whether they are abstract or
concrete.

Since a #-generated function my end up in an extra anonymous module, we
have to make sure that treatAbstractly works correctly (ignoring this
extra module).  Hence the fix for isNoName.