[ warning ] Highlighting positivity issues
* killrange when solving metas

In order to avoid positivity issues to be reported with the range
of the term that made solving the meta possible rather than where
the meta actually originated from. As a consequence a few golden
valus are changed.

* Keeping the record's tele's ranges

In order to be able to report positivity issues, we need to keep
the ranges in the record's telescope.