[ new ] Anonymous Declarations
This commit introduces anonymous declarations which will be
scope-checked and type-checked but won't be polluting the
scope (i.e. it is impossible to mention them in the module
they are defined in and they are not exported). They only
have to conform to some simple conditions:

* They are all named "_"
* They cannot bind arguments on the LHS (but one can use
  lambda-abstractions on the RHS)
* They cannot be defined by recursion.

I guess this (re)-closes #1465