[ emacs-mode ] solveAll can normalise solutions
When writing examples of the value a function produces
when it is normalised, it is a bit frustrating to have
`C-c C-s` generate an utterly useless solution.

E.g. the goal in the following code snippet will be
solved using `reverse (0 ∷ 1 ∷ 2 ∷ [])` when we would
have wanted `2 ∷ 1 ∷ 0 ∷ []` to be used instead.

------------------------------------
_ : reverse (0 ∷ 1 ∷ 2 ∷ []) = ?
_ = refl
------------------------------------

Following this commit, `C-c C-s` keeps its default behaviour
however using the `C-u` prefix will yield a 'Simplified'
version of the solutions and using `C-u C-u` will yield the
fully 'Normalised' ones.