[ Fix #2268 ] Highlighting of comments & pragmas
Using "font-lock-syntactic-face-function" we only attach a face to
comments wich start either with "--" followed by a space or a newline
or ones which start with "{-" followed by anything but "#".

This way we leave pragmas (they start with "{-#") and options (they
start with "--[a-z]") alone whilst statically highlighing all (?)
the comments.