[ #1343 ] Calling `(C-u)* C-c C-s` from a goal
Instead of solving *all* the goals which are already instantiated
internally, one may want to only solve the current one. It is now
possible. The fine control over how normalised the solution should
be via (C-u)* prefixes is also supported (by default the solution
is returned as is).

This closes #1343.