[ fixed #2303 ] as-bindings leaking from parent clause to with-clause
This required a change to stripWithClausePatterns which previously
dropped module parameter arguments in the parent clause. This worked
because of the as-binding leaking, but with that fixed stripWithClausePatterns
has to process the module parameter arguments as well.

This in turn revealed a bug in with the inherited dot-patterns generated
by stripWithClausePatterns. These are expected to live in the parent clause
lhs context, but were generated to live in the parent function argument
telescope.