[ agda-ghc-names ] moved to its own repo (agda/agda-ghc-names)
The main reason for doing this is so we can get rid of the haskell-src-exts
dependency.

I think separating agda-ghc-names from the main repo ok since it's
a very specialised tool with (I expect) very few users, and it doesn't
depend heavily on the internals of Agda so it's not critical that you use
matching versions.

Fixes #2121.