[ fixed #2018 ] don't rename variables when adding them to the context
... but make sure to pick fresh names when the system is choosing the name.

This means we have to pick fresh names for clauses generated by make case
in the AbstractToConcrete translation instead of relying on addContext to do it
for us. That makes more sense anyway since the old implementation used the
context of the left-hand side before pattern matching when picking fresh names.