Fixed 2046 by an agreeably dirty hack (no clean up yet)
We first cluster the constraints by overlap in size metas.  This used to
happen on the size constraints in the SizeExp syntax, but now we do it
on the constraints in Internal syntax.

The we pick for each cluster a constraint with the longest context and
try to convert the other constraints to that context.

To match the contexts of constraints for size solving, we compare CtxIds,
but if this fails, names of bound variables (douh!).

This is highly unsafe and can probably easily be broken with shadowing.

What we really need is a proper size constraint solver that can deal in
generality with quantified constraints (see issue #300).