More detailed benchmarking for operator parsing.
I noticed that the result of parsing a pattern is often ambiguous. The
reason is that when we parse a pattern the non-constructor operator name
parts in scope can be rebound. For instance, the following pattern can
be parsed in two distinct ways:

  infixl 4 _+_+_ _+_

  postulate
    _+_ : Set

  _+_+_ : Set → Set → Set → Set₁
  _ + _ + _ = Set

However, only one of these parses is well-formed given the information
that the symbol that is defined is _+_+_.

For this reason I wanted to know how much of the time spent on parsing
operators is spent on parsing patterns. For the standard library it
seems as if this fraction is roughly 20% (with my current setup).