Fixed #2354: need to revisit constraints after size solver
Interaction between size solver, constraint solver, and instance search.

The instance search fails initially because of a non rigid meta.
Before freezing, there is another attempt.
The instance search succeeds and instantiates the meta for M.
Then, a size constraint is solved which had blocked a term.
Finally the blocking constraint can be resolved.

Thus, we need to
* wake constraints
* solve size constraints
* wake constraints again.
In principle, there could be many alternations needed!?

This change caused reordered printing of left-over constraints in
the *.err files.