[ pragma ] Allowing --safe in the OPTIONS pragma
In order to make it easier for users to state explicitly that their
modules only use the safe subset of the language, it is now possible
to use `--safe` in an `OPTIONS` pragma.

Incompatibilities between `--safe` and `postulate`, `primTrustMe`,
`(NON_)TERMINATING`, `--no-termination-check`, `--no-positivity-check`,
`--rewriting`, etc. give rise to non-fatal errors.

This effort is related to https://github.com/agda/agda-stdlib/pull/122
which aims to make it easier to program in the safe subset of the
language when using the standard library.

This closes Issue #2442.