[ major refactor ] change the representation of compiler pragmas
During type checking, compiler pragmas for definitions are now stored as
plain strings. For instance, a pragma

```
{-# COMPILED_DATA Bool Bool False True #-}
```

is represented internally as a GHC backend pragma `"= data Bool (False | True)"`.
This means that the backends have to do more work parsing pragmas than before.

The reason for the change is to enable backend-agnostic pragma syntax, where
the above pragma would be written

```
{-# COMPILE GHC Bool = data Bool (False | True) #-}
```