[ #2492 ] Folding ℕs during Abstract to Concrete
We turn `suc^n zero` into `Lit n` when converting from Abstract
to Concrete. The most obvious benefit is that Auto doesn't spit
out massive terms when it has built a big natural number (however
it still takes a long time: internally Auto does generate the big
chain of `suc`s and we need to go through it.

There is another opportunity to call `tryToRecoverNatural` in
`toConcrete` in the `Con` case:

    toConcrete (Con (AmbQ (x:_))) = Ident <$> toConcrete x

However that turns the warning about `{-# BUILTIN ZERO zero #-}`
into a weird warning about `{-# BUILTIN ZERO 0 #-}` so I figured
it was not worth the trouble.