[ refactor ] typeclass magic replacing valu[0..25]
Using some typeclass magic we can define a function valuN
which covers the same ground as all of the valu0 to valu25.

Because the concrete instances of `valuN` are generated by
typeclass resolution, they are all built and inlined at
compile time.

Looking only at deserialisation for the standard library,
this leads to a ~1.3s speed-up (about 9%) on my machine.
The next step is to try to do the same for serialisation.