[ refactor ] typeclass magic for icode(')[0..25]
Using the same typeclass magic as for valuN, we define
the generic functions icodeN and icodeN'. Rather than
taking a Proxy as an argument, they take the constructor
they encode the arguments of (but do not use it i.e. a
type-annotated `undefined` can be used in some cases).

The new icodes are *more typed* versions of the previous
serialisation combinators which can help catch some bugs
such as the mismatch between

    icod_ (OldBuiltin a b) = icode1 1  a

and

    valu [1, a, b]     = valuN OldBuiltin a b

in the `EmbPrj Warning` instance which is fixed by this
commit.

Looking at the standard library, the new serialisation
phase seems to behave marginally better.

As for valuN, this is incompatible with 7.6.3 (cc #1972)