[ refactor ] Getting rid of the value[1..25] too
Same typeclass magic as for valuN and icodeN,
Same marginally better results on stdlib,

This should mark the end of this refactoring effort. Overall
Typechecking.Serialise.Base will have shrunk by about 550 lines.

This is also incompatible with 7.6.3 so cc 1972 in case we decide
to revert some commit for the next release.