[ refactor ] Reader monad + Functional dependency
Given that two of the arguments were constant throughout the
recursive calls, we might as well just put them in a reader
monad.

We also add UndecidableInstances to get ghc to accept that
there is indeed a functional dependency which makes it possible
to determine `o` from `t` and `u` in `Replace o t u`.
The problematic case was `Replace o (Abs t) (Abs u)` which is
forced by the premise `Replace o t u`.