[ new ] Support for Agda.Builtin.Strict (#157)
* [ new ] Support for Builtin.Strict

Agda.Builtin.Strict was introduced in February 2016 but we
somehow missed it and it's not part of the standard library.
This fixes this issue.

Also, I've taken the opportunity to introduce ticked version
of various apply functions which, just like the ticked
composition operator, expect their function *not* to be a
dependent function. This makes type-inference a lot easier
and is useful to make unsolved metas go away!