[ cubical ] A.T.Records.typeElims now accepts a Path type for Apply elims.
This fixes a regression introduced by the a07cee9 merge of the refactor at dea698f.
To be fair it might be better to modify the syntax of patterns so that we can
generate IApply, as opposed to Apply, when eliminating a Path type.
Right now the VarP constructor does not have the information necessary.