Fixed #2703: getConType crashes on partially applied constructors with indices
The fix works around the missing laziness of substitution into Defs.
Doing the strengthening when it is 100% safe fixes the issue.

From the issue discussion on github:

I don't understand the error. All we want from the strengthened type is
the datatype parameter, which should be well-scoped. We ignore the
indices which have the impossible variables.
...
I think I located the problem: It is the way we apply substitutions to `Def`s:
```haskell
instance Subst Term Term where
  applySubst rho t    = case t of
    Var i es    -> lookupS rho i `applyE` applySubst rho es
    ...
    Def f es    -> defApp f [] $ applySubst rho es

-- | @defApp f us vs@ applies @Def f us@ to further arguments @vs@,
--   eliminating top projection redexes.
--   If @us@ is not empty, we cannot have a projection redex, since
--   the record argument is the first one.
defApp :: QName -> Elims -> Elims -> Term
defApp f [] (Apply a : es) | Just v <- canProject f (unArg a)
  = argToDontCare v `applyE` es
defApp f es0 es = Def f $ es0 ++ es

-- | If $v$ is a record value, @canProject f v@
--   returns its field @f@.
canProject :: QName -> Term -> Maybe (Arg Term)
canProject f v =
  case ignoreSharing v of
    (Con (ConHead _ _ fs) _ vs) -> do
      i <- List.elemIndex f fs
      headMaybe (drop i vs)
    _ -> Nothing
```
In the particular case, `applySubst (StrengthenS...) (Def f var0)` will check whether `strengthen var0` is a record constructor with field `f`, and crashes.

Thus, my assumption about the laziness of substitution is wrong.

The workaround of my last comment fixes the present issue.

However, to get the lazy behavior I imagined, we would have to store in `Def` whether the `QName` can be a record projection or not.  Anyway, I somehow feel that using `Def` for record and datatypes is not the best factoring.  (See also the numerous calls to `allApplyElims` we have as a consequence of the current structure.)