[ flat ] Changed the modalities schema, to support spatial type theory's flat.
We revert back to some of the original behaviour for Irrelevance/NonStrict.

CoShape is now a flat/necessity/global sections/limit modality.
 The UI writes flat or ♭ for it.
Sharp   now annotates completely unusable variables
	, semantically could be constantly the terminal object.
 The UI writes top or ⊤ for it.

(note that Sharp is no longer intended to be a Sharp from cohesion)

In cases like composition of CoShape and Irrelevance or other pairings
for which we don't have a model together we produce Sharp, not necessarily
sound but close enough for now, hopefully.