[ forcing ] code to rewrite clauses using forced variables (disabled)
Instead of using the forced variable it's always possible
to find a dot pattern that can be turned into an actual
pattern binding the variable instead.

The translation takes place in the clause compiler,
before generating a case tree, except it's disabled
for now, since we're not yet checking that we don't
rebind a variable to a binding with the wrong relevance.
The plan is to have the type checker take care of this.