[ #2474 ] Support for warning highlighting
Previously warnings were not displayed in the LaTeX and HTML
backends as well as in imported modules in emacs.

Instead of trying to reconstruct warning highlighting a posteriori,
we generate it before building the interface, merge it with the rest
of the syntax highlighting info and store it.

Turning on the support for `otherAspects` in LaTeX has revealed
a missing `\newcommand` definition for `AgdaCatchallCaluse` which
has been added (it does not do anything at the moment).

fixes #2474
closes #2895