[ #2474 ] agda.sty + fix golden value + test
AgdaCatchallClause was not defined in agda.sty and the test suite
failed because of that. This is now ok.

The modified golden value shows a caveat to the approach I took:
it produces commands that span at most a token even when things
ought to be highlighted as a block.

In the future it may be worth investigating whether it's possible
to highlight entire blocks rather than isolated tokens. To do so
we'd need to:

* group various consecutive tokens based on their shared OtherAspects
* figure out the nesting order for these OtherAspects as they may not
  all span the same tokens (e.g. a CatchallClause would span the whole
  LHS when an unsolved constraint may appear in only one of the args)
* see whether that interacts with the alignment machinery