Lots of changes to the graph library.
I noticed that nodes without incoming and outgoing edges were not
handled properly by the transposition function (they were thrown
away). In order to make it easier for me to understand the code I
decided to enforce a new invariant: every target node must be present
in the graph as a source node (possibly without outgoing edges). In
order to make sense of this graphs are now required to be homogeneous,
with source and target nodes having the same type.

This means that some functions (notably insertWith) are now
implemented a bit differently, in order to establish or preserve the
invariant. This could affect other parts of Agda, but given that the
changes make graphs larger I suspect that they will not lead to more
programs being type-correct, and the test suite does not indicate any
problems.

Quick testing suggests that performance is more or less
unaffected. (The time required to process Issue1560.agda has perhaps
increased by something like 1%.)

Other changes:
* The transpose function (which is now exported) no longer removes
  nodes without incoming or outgoing edges.
* The sourceNodes function no longer returns nodes without outgoing
  edges, and the srcNodes field of the Nodes structure is treated
  analogously.
* Some functions no longer remove source nodes without outgoing
  edges: clean, composeWith, filterEdges, removeNode.
* The Pretty instance now prints nodes without outgoing edges.
* The Show instance is no longer derived. The Graph type is abstract,
  so the new Show instance yields a graph expression that can be used
  to reconstruct the graph (just like the Show instance for
  Data.Map.Map).
* The completeIter function is now exported.
* The new functions fromNodeSet, removeNodes and isolatedNodes are
  also exported.
* The functions fromList and fromListWith have been renamed to
  fromEdges and fromEdgesWith.
* The toList function has been removed (it was basically a copy of
  edges).
* Various changes to the documentation.