[ fixed #2883 ] wrong parameters to pattern lambda
Gets rid of the hidden/not hidden parameters in ExtLamInfo.
These don't make any sense in the presence of parameter
refinement. Instead we create a fresh module to track the
parameters of the pattern lambda. It would be nice if the lambda
could live in this module, but at the moment the module is
decided at scope checking and this looks difficult to change.