[ printing ] print metas with substitutions rather than as globals
E.g. we would print "?0 (x = a) (y = b)" rather than "?0 a b" if "?0"
is created with "x" and "y" in scope.

It also tries to clean up redundant parts of the subsitution by using
checkpoints to line up the current context and the one the meta was
created in. This is to avoid printing e.g. "(x = x)" in "?0 (x = x)".

It appears though that different definitions inside a parametrized
module do not have a common checkpoint corresponding to the module
parameters, so we get the whole substitution printed out in those
cases.