[ resp ] Include the interaction callback in TCM
The datatype `Response` are made and assembled inside TCM,
before being converted into Lisp or JSON in the IO monad.

  +-----------------------------------------------------+
  | IO                                                  |
  |  +----------------+                  lispify        |
  |  | TCM            |                +--------> Lisp  |
  |  |  Interaction ----> Response ----|                |
  |  |                |                +--------> JSON  |
  |  +----------------+                  jsonify        |
  |                                                     |
  +-----------------------------------------------------+

However, many of the information carried by `Response` are formatted as
Strings in TCM (for Emacs consumpiton). This makes it diffucult for
other editors to analyse and reconstruct those information.

And it seems like those strings can ONLY be assembled and formatted in
TCM because there are too many states that has to be accessed on demand.
(see `PrettyTCM` for example).

One solution is to specialise `Reponse`. Let everyone decide what they
want to encode inside TCM. But I don't think this is a good idea :|

[1]
  +-----------------------------------------------------+
  | IO                                                  |
  |  +----------------+                                 |
  |  | TCM            |                 lispify         |
  |  |  Interaction +---> LispResponse +--------> Lisp  |
  |  |              | |                                 |
  |  |              | |                 jsonify         |
  |  |              +---> JSONResponse +--------> JSON  |
  |  |                |                                 |
  |  +----------------+                                 |
  |                                                     |
  +-----------------------------------------------------+

Another way is to simply extend the scope of TCM to cover the formatting
process like this:

[2]
  +-----------------------------------------------------+
  | TCM                                                 |
  |                                      lispify        |
  |                                    +--------> Lisp  |
  |     Interaction ----> Response ----|                |
  |                                    +--------> JSON  |
  |                                      jsonify        |
  |                                                     |
  +-----------------------------------------------------+

This PR implements the latter solution because I think it looks more
reasonable, and it's also surprisingly easy to implement.
Only 5 lines of code is modified and the impact on the codebase is
minimum.

I guess the reason why the lispify process was done in the IO monad in
the first place is that, we want to prevent it from messing with the
states inside TCM ???
I think perhaps we can prevent that from happening but also being able
to access those states by wrapping it in another reader-only monad,
but I'm not sure if it's worth it.

  +-------------------+---------------------------------+
  | TCM               | TCM' (read-only)                |
  |                   |                  lispify        |
  |                   |                +--------> Lisp  |
  |     Interaction ----> Response ----|                |
  |                   |                +--------> JSON  |
  |                   |                  jsonify        |
  |                   |                                 |
  +-----------------------+-----------------------------+