[ fix ] be more mindful of dependencies
And throw an error when we couldn't find a coping strategy for all
of the lone signatures at hand.

I only handle function definitions at the moment but it should be
possible to handle datatypes and record types too. We should be able
to compute the type of the axiom from the list of parameters and the
level the declaration lives at.