[ warning ] AbsurdPatternRequiresNoRHS
There is no reason to raise an error if the user gives a RHS despite
the presence of an absurd pattern: we can safely ignore it!

Took the opportunity to rename `ReachabilityProblem` to `Deadcode`
which is more general. In the future this aspect can also be used
for empty pragmas, unused let / where bindings, empty blocks, fixity
declarations without a corresponding definition, unused imports, etc.