[ fix #3199 ] Distinguish imported & local user warnings
The problem with putting (recursively) imported user warnings in the
interface is that they refer to identifiers defined in modules which
are not themselves directly imported. And Agda doesn't know how to
translate these module names to paths during serialization!

We now only put the user warnings defined locally in the interface.